feat(auth): make server actions specific to the logged in user
after this, if a user tries to delete/edit/change status of ticket they
do not own will result in an unauthorized error

+++ src/features/tickets/actions/delete-ticket.ts
@@ -3,12 +3,23 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 
 export const deleteTicket = async (ticketId: string) => {
+  const { user } = await getAuthOrRedirect();
   try {
+    const ticket = await prisma.ticket.findUnique({ where: { id: ticketId } });
+
+    if (!ticket || !isOwner({ authUser: user, entity: ticket })) {
+      return toActionState({ status: "ERROR", message: "Not authorized" });
+    }
     await prisma.ticket.delete({ where: { id: ticketId } });
   } catch (error) {
     return fromErrorToActionState({ error });
